refactor(examples-switcher): boundary as booleans, effects into hooks, one unit for the height
`shown === 'block'` made the panel height depend on which CSS keyword the
visibility token happens to use — switching DESKTOP_ONLY to 'flex', the form
used elsewhere in this repo, would silently drop every desktop slot back to
PANEL_HEIGHT. The boundary is now DESKTOP_SLOTS booleans in styles.js, with
MOBILE_ONLY, DESKTOP_ONLY and onDesktop() all derived from it.

- Both effects move to use-tabs-height.js and use-scroll-cues.js with their
  bodies intact, taking the component from 357 to 285 lines and leaving one
  ResizeObserver block per file instead of two in one. getScrollCues and
  TAB_GAP_PX move with their only callers.
- The measured height was carried as viewportHeight (number) and tabsHeight
  (string) at once, with TabList re-appending px inside the styled function.
  TabList takes $height, the px string, and the `visibleTabs &&` in the
  derivation goes: the hook already nulls the height when visibleTabs is unset.
