+++ src/components/patterns/ExamplesSwitcher/styles.js
@@ -1,8 +1,13 @@
 import { colors, theme } from 'theme'
 import { css } from 'styled-components'
 
-export const MOBILE_ONLY = ['block', 'block', 'none', 'none']
-export const DESKTOP_ONLY = ['none', 'none', 'block', 'block']
+const DESKTOP_SLOTS = [false, false, true, true]
+
+export const MOBILE_ONLY = DESKTOP_SLOTS.map(on => (on ? 'none' : 'block'))
+export const DESKTOP_ONLY = DESKTOP_SLOTS.map(on => (on ? 'block' : 'none'))
+
+export const onDesktop = (desktopValue, mobileValues) =>
+  DESKTOP_SLOTS.map((on, index) => (on ? desktopValue : mobileValues[index]))
 
 export const SECONDARY_FOCUS_RING = css`
   outline: 2px solid ${colors.secondary};
